==> notebooks/networks_preparation/2-binarize_weighted_networks.ipynb <==
import pandas as pd
import numpy as np
import os

weighted_networks_dir = "../data/networks/all/weighted/"
binarized_weighred_networks_dir = "../data/networks/all/binarized_weighted/"

os.makedirs(binarized_weighred_networks_dir, exist_ok=True)

for path in os.listdir(weighted_networks_dir):
    if path.endswith(".csv"):
        network = pd.read_csv(f"{weighted_networks_dir}{path}")
        network.set_index("Plant", inplace=True)
        network = (network > 0).astype(np.int16)
        network.reset_index(inplace=True)
        network.to_csv(f"{binarized_weighred_networks_dir}{path}", index=False)
