==> notebooks/networks_preparation/4-extract_networks_locations.ipynb <==
remotes::install_github("ropensci/rmangal")

library(tidyverse)
library(magrittr)
library(dplyr)
library(sf)
library(rmangal)
library(raster)

networks_metadata_path = "../../data/networks/all/networks_metadata.csv"

rmangal_observations_data_path = "../../data/networks/metadata/parsed_rmangal_networks_metadata.csv"
rmangal_metadata_path = "../../data/networks/metadata/all_rmangal_networks.csv"

ecoregion_polygon_data_path1 = "../../data/networks/metadata/polygon/WWF/wwf_terr_ecos.shp"
ecoregion_polygon_data_path2 = "../../data/networks/metadata/polygon/Ecoregions2017/Ecoregions2017.shp"

temp_input_for_bioclim_path = "../../data/networks/all/geo_avail_networks_metadata.csv"
temp_output_for_bioclim_path = "../../data/networks/all/geo_extracted_networks_metadata.csv"
bioclim_dir = "../../data/networks/metadata/bioclimate/"
richness_data_path = "../../data/networks/metadata/Richness_2021.tif"
rmangal_geodata_path = "../../data/networks/rmangal_networks_geodata.csv"

# Mangal

if (!file.exists(rmangal_metadata_path))
{
    pollination_interactions<- search_interactions(type = "mutualism")
    network_ids = unique(pollination_interactions$network_id)
    length(network_ids)
    
    all_networks_data = data.frame()
    network_metadata_dfs = list()
    num_avail = 0
    for (i in 1:length(network_ids)) #length(pollination_networks)
    {
      if (i %% 10 == 0) 
      {
        print(i)
        print(num_avail)
      }
      networks = search_networks(list(id=nth(network_ids,i)))  %>% get_collection
      network = networks[[1]]
      network_data = network$network[c("geom_lat", "geom_type", "geom_lon")]
      network_ref_data = network$reference
      network_data = bind_cols(network_data, network_ref_data)
      network_data[["network_id"]] = nth(network_ids,i)
      network_metadata_dfs[[i]] = network_data
      if (! is.na(network_data$geom_lon))
      {
        num_avail = num_avail + 1
      }
    df = ldply(network_metadata_dfs, rbind)
    df <- apply(df,2,as.character)
    write.csv(df,rmangal_metadata_path)
    }
}

rmangal_metadata = read.csv(rmangal_metadata_path)

head(rmangal_metadata[is.na(rmangal_metadata$geom_lat), ][c("network_id", "doi", "paper_url", "geom_lon", "geom_lat")])

# merge rmangal_metadata to networks_metadata

# Web of life

base_url <- "https://www.web-of-life.es/"
networks_metadata = read.csv(networks_metadata_path)
non_wol_networks_metadata = networks_metadata[networks_metadata$source != "web_of_life", ] 
wol_networks_metadata = networks_metadata[networks_metadata$source == "web_of_life", ][c('network_type','network_index','path','source','is_legal','is_resolved','processed_path','is_parsed_legal')]                                                                                      
all_nw_info <- read.csv(paste0(base_url,"get_network_info.php"))

get_name <- function(path)
{
    path_name = basename(path)
    name = unlist(strsplit(path_name, ".", fixed=TRUE))[1]
    return (name)
}

wol_networks_metadata$network_name = unlist(lapply(wol_networks_metadata$path,get_name))

all_nw_info$Latitude <- all_nw_info$latitude
all_nw_info$Longitude <- all_nw_info$longitude

drops <- c("network_type","number_components", "is_weighted", "cell_values_description", "abundance_description", "latitude", "longitude")
all_nw_info = all_nw_info[ , !(names(all_nw_info) %in% drops)]

colnames(all_nw_info)

wol_networks_metadata <- left_join(wol_networks_metadata, all_nw_info, by = join_by(network_name == network_name))

rest_networks_metdata =  networks_metadata[networks_metadata$source != "web_of_life", ]

networks_metadata = bind_rows(wol_networks_metadata, rest_networks_metdata)

write.csv(networks_metadata, networks_metadata_path)
