==> notebooks/networks_preparation/7-cluster_ecoregions_by_koppen_classification.ipynb <==
library(kgc)
library(dplyr)
library(tidyr)

network_metadata_path = "../../data/networks/networks_metadata.csv"

network_metadata = read.csv(network_metadata_path)
relevant_network_data = drop_na(network_metadata, any_of(c("Longitude", "Latitude")))
relevant_network_data = relevant_network_data[, c("network_type", "network_index", "Longitude", "Latitude")]

relevant_network_data = data.frame(relevant_network_data, rndCoord.lat = RoundCoordinates(relevant_network_data$Latitude), rndCoord.lon = RoundCoordinates(relevant_network_data$Longitude))

relevant_network_data = data.frame(relevant_network_data,ClimateZ=LookupCZ(relevant_network_data))

relevant_network_data <- data.frame(relevant_network_data, CZUncertainty(relevant_network_data))

relevant_network_data[relevant_network_data$ClimateZ == "Climate Zone info missing", c("ClimateZ")] = NA

# categorization is devised from http://koeppen-geiger.vu-wien.ac.at/present.htm
ClimateZ = c(equatorial = c("Af", "Am", "As", "Aw"),
             arid = c("BSh", "BSk", "BWh", "BWk"),
             warm_temprate = c("Cfc", "Cfb", "Csc", "Csb", "Cfa", "Csa", "Cwa", "Cwb", "Cwc"),
             boreal = c("Dwc", "Dwb", "Dsa", "Dsb", "Dfb", "Dfc", "Dsc", "Dwa", "Dfa", "Dwd", "Dfd"),
             polar = c("EF", "ET"))
clim_cat_data = data.frame(ClimateZ)
clim_cat_data["climate_category"] = rownames(clim_cat_data)
clim_cat_data$climate_category = gsub("[0-9]*","",clim_cat_data$climate_category)
rownames(clim_cat_data) = NULL

relevant_network_data <- merge(relevant_network_data, clim_cat_data, by.x = "ClimateZ", 
             by.y = "ClimateZ", all.x = TRUE, all.y = FALSE)

network_metadata = merge(network_metadata, relevant_network_data[, c("network_type", "network_index", "ClimateZ", "climate_category")], by.x = c("network_type", "network_index"), 
             by.y = c("network_type", "network_index"), all.x = TRUE, all.y = FALSE)

head(drop_na(network_metadata, any_of(c("Latitude")))[, c("network_type", "network_index", "Latitude", "Longitude", "climate_category")])

unique(network_metadata$climate_category)

debug = network_metadata[is.na(network_metadata$climate_category), c("EcoRegion_name", "ClimateZ")] # "Longitude", "Latitude", 
debug = unique(debug)
debug

write.csv(network_metadata, network_metadata_path)
