==> notebooks/networks_preparation/8-remove_duplicated_networks.ipynb <==
import numpy as np
import pandas as pd
import os
pd.set_option('display.max_colwidth', None)

networks_metadata_path = "../../data/networks/networks_metadata.csv"
networks_dir = "../../data/networks/all/"

networks_metadata = pd.read_csv(networks_metadata_path)
networks_metadata = networks_metadata.rename(columns={"duplicated_network": "mangal_duplicated_network"})

# report allegendly duplicated networks from mangal based on the data url

debug = networks_metadata.loc[(networks_metadata.mangal_duplicated_network == True) & (networks_metadata.data_url.str.contains("interactionweb", na=False)) | (networks_metadata.data_url.str.contains("web-of-life", na=False))][["processed_path", "Longitude", "Latitude", "data_url", "num_plants", "num_pollinators"]].sort_values(["num_plants", "num_pollinators"]).drop_duplicates()
debug

# detect duplications by first selecting suspects of duplicated networks based on networl size and then directly checking if the networks are indentical within each group of suspects

def is_equal_net(net1, net2):
    same_plants = set(net1.Plant.str.lower()) == set(net2.Plant.str.lower())
    same_pollinators = set([s.lower() for s in net1.columns if s != "Plant"]) == set([s.lower() for s in net2.columns if s != "Plant"])
    try:
        same_inter = np.all(net1.set_index("Plant").to_numpy() == net2.set_index("Plant"))
    except:
        same_inter = False
    return same_plants & same_pollinators & same_inter
    
    
def get_duplicated_clusters(network_paths):
    path_to_net = {path:pd.read_csv(path) for path in network_paths}
    for p in path_to_net:
        if "Unnamed: 0" in path_to_net[p]:
            path_to_net[p] = path_to_net[p].drop(["Unnamed: 0"], axis=1)
    dups = []
    for i in range(len(network_paths)):
        for j in range(i+1, len(network_paths)):
            if is_equal_net(path_to_net[network_paths[i]], path_to_net[network_paths[j]]):
                dups.append((network_paths[i], network_paths[j]))
    return dups
                


debug = networks_metadata.groupby(["network_type", "num_plants", "num_pollinators"])[["network_index", "path", "processed_path"]].agg({"network_index": lambda n: len(n),
                                                                                                                                                         "path": lambda n: list(n),
                                                                                                                                                         "processed_path": lambda n: list(n)}).reset_index()
debug["from_oik07303"] = debug.path.astype(str).str.contains("oik.07303", na=False)
debug["duplicated_networks"] = debug.processed_path.apply(get_duplicated_clusters)

debug.loc[debug.duplicated_networks.astype(str) != "[]"]
paths_to_remove = debug.loc[debug.duplicated_networks.astype(str) != "[]"].path.apply(lambda x: x[1:]).tolist()
paths_to_remove = [item for sublist in paths_to_remove for item in sublist]
paths_to_remove

networks_metadata = networks_metadata.loc[~networks_metadata.path.isin(paths_to_remove)]

print(f"# remaining networks per type: ")
networks_metadata.groupby("network_type")["network_index"].count()

print(f"# remaining unique networks of all original types = {networks_metadata.loc[networks_metadata.network_type != 'binarized_weighted'].shape[0]:,}")

networks_metadata.to_csv(networks_metadata_path, index=False)

# less duplicates were detected compared to the ones reported as duplicates from mangal, so allegedly duplicated from mangal will be checked manuall

mangal_dups = networks_metadata.loc[(networks_metadata.mangal_duplicated_network == True) & (networks_metadata.data_url.str.contains("web-of-life", na=False)) & (networks_metadata.network_type == "weighted")][["processed_path", "EcoRegion_name", "num_plants", "num_pollinators"]].sort_values(["num_plants", "num_pollinators"]).drop_duplicates()
mangal_dups.iloc[:30]

wof_nets = networks_metadata.loc[(networks_metadata.source == "web_of_life") & (networks_metadata.network_type == "weighted")][["processed_path", "EcoRegion_name", "num_plants", "num_pollinators"]].sort_values(["num_plants", "num_pollinators"]).drop_duplicates()
wof_nets.iloc[:30]
